Move ShowDialog and Write logic from Debug to DebugProvider
- Also removing delegates
